chore: use workspaces
