--- packages/kit-tamagui/vitest.config.mjs
@@ -1,5 +1,0 @@
-export default {
-  esbuild: {
-    format: 'cjs',
-  },
-}
