+++ packages/kit-tamagui/vitest.mjs
@@ -1,0 +1,3 @@
+import { react } from '../../test/config.mjs'
+
+export default react
