+++ packages/server/src/__tests__/http-server.test.ts
@@ -1,6 +1,3 @@
-/**
- * @vitest-environment node
- */
 import { afterAll, describe, expect, test } from 'vitest'
 
 import { createHTTPServer } from '../http-server.js'
