+++ packages/server/vitest.config.mjs
@@ -1,0 +1,3 @@
+import { node } from '../../test/config.mjs'
+
+export default node
