+++ vitest.workspace.js
@@ -1,0 +1,1 @@
+export default ['./packages/**/*.config.mjs']
